Split http module into focused submodules
